Alert Message when no information provided
Created an alert system when user click on search without providing any information Set a logic if statement if text is blank in the Search.js. initailaize a state for alert and set a timeout to clear the alert screen. Created Alert.js to print the message alert to the screen and making sure that the alert state is not null before displaying the said alert.